Add Kalman strategies to reduce initial noise
The first strategy consists on extrapolating points, add them to the origin and then run the kalman filter. Result smoothed points are then trimed at the beggining, by the same amount of points that were extrapolated. This strategy is fast, but it changes the properties of the track, as the points are not extremely correct.

The second strategy consists on runnning the kalman filter two time: one time for the whole track, the second time the kalman filter is applied to a reversed subset of the track. Points are then removed from the beggining of each, and they are merged. This stratefy is slower, as requires the kalman filter, and the *em* method to be applied twice, but it is more truthful to the properties of track.